Replaced simd in scoring lookup because it was slow.

+++ src/alignment_cmdline.c
@@ -408,7 +408,7 @@ void align_from_query_and_db(const char *query_path, const char *db_path, scorin
     char *query_fasta = query_read.name.b;
     size_t query_seq_len = query_read.seq.end;
     // characters are converted into indexes for table lookup
-    int32_t * query_indexes = aligned_alloc(32, query_seq_len * sizeof(int32_t));
+    int8_t * query_indexes = aligned_alloc(32, query_seq_len * sizeof(int8_t));
 
     // Replace unknown characters in query with an X
     for(i = 0; i < query_seq_len; i++) {
@@ -429,7 +429,7 @@ void align_from_query_and_db(const char *query_path, const char *db_path, scorin
         aligners[i] = NULL;
     }
 
-    int32_t *db_seq_index_vec_batch = NULL;
+    int8_t *db_seq_index_vec_batch = NULL;
     char **db_seq_vec_batch = NULL;
     char **db_fasta_vec_batch = NULL;
     size_t vec_elem_cnt = 0;
@@ -454,7 +454,7 @@ void align_from_query_and_db(const char *query_path, const char *db_path, scorin
             // since db is sorted from longest to shortest, first item in batch
             // will always be the longest.
             max_seq_len_in_vec = seq_b_len;
-            db_seq_index_vec_batch = aligned_alloc(32, max_seq_len_in_vec * VECTOR_SIZE * sizeof(int32_t));
+            db_seq_index_vec_batch = aligned_alloc(32, max_seq_len_in_vec * VECTOR_SIZE * sizeof(int8_t));
             db_seq_vec_batch = malloc(sizeof(char *) * VECTOR_SIZE);
             db_fasta_vec_batch = malloc(sizeof(char *) * VECTOR_SIZE);
         }
